+++ packages/storybook-solid/src/public-types.ts
@@ -12,6 +12,7 @@ import type {
     StoryContext as GenericStoryContext,
     StrictArgs,
 } from '@storybook/types';
+
 import type { Component as ComponentType, ComponentProps } from 'solid-js';
 import type { SetOptional, Simplify } from 'type-fest';
 
@@ -21,70 +22,57 @@ export type { SolidRenderer };
 /**
  * Metadata to configure the stories for a component.
  *
- * @see [Default export](https://storybook.js.org/docs/formats/component-story-format/#default-export)
+ * @see [Default export](https://storybook.js.org/docs/api/csf#default-export)
  */
-export type Meta<TCmpOrArgs = Args> =
-  TCmpOrArgs extends ComponentType<any>
-      ? ComponentAnnotations<SolidRenderer, ComponentProps<TCmpOrArgs>>
-      : ComponentAnnotations<SolidRenderer, TCmpOrArgs>;
+export type Meta<TCmpOrArgs = Args> = [TCmpOrArgs] extends [ComponentType<any>]
+    ? ComponentAnnotations<SolidRenderer, ComponentProps<TCmpOrArgs>>
+    : ComponentAnnotations<SolidRenderer, TCmpOrArgs>;
 
 /**
  * Story function that represents a CSFv2 component example.
  *
- * @see [Named Story exports](https://storybook.js.org/docs/formats/component-story-format/#named-story-exports)
+ * @see [Named Story exports](https://storybook.js.org/docs/api/csf#named-story-exports)
  */
-export type StoryFn<TCmpOrArgs = Args> =
-  TCmpOrArgs extends ComponentType<any>
-      ? AnnotatedStoryFn<SolidRenderer, ComponentProps<TCmpOrArgs>>
-      : AnnotatedStoryFn<SolidRenderer, TCmpOrArgs>;
+export type StoryFn<TCmpOrArgs = Args> = [TCmpOrArgs] extends [ComponentType<any>]
+    ? AnnotatedStoryFn<SolidRenderer, ComponentProps<TCmpOrArgs>>
+    : AnnotatedStoryFn<SolidRenderer, TCmpOrArgs>;
 
 /**
- * Story function that represents a CSFv3 component example.
+ * Story object that represents a CSFv3 component example.
  *
- * @see [Named Story exports](https://storybook.js.org/docs/formats/component-story-format/#named-story-exports)
+ * @see [Named Story exports](https://storybook.js.org/docs/api/csf#named-story-exports)
  */
-export type StoryObj<TMetaOrCmpOrArgs = Args> = TMetaOrCmpOrArgs extends {
-    render?: ArgsStoryFn<SolidRenderer, any>;
-    component?: infer Component;
-    args?: infer DefaultArgs;
-}
+export type StoryObj<TMetaOrCmpOrArgs = Args> = [TMetaOrCmpOrArgs] extends [
+    {
+        render?: ArgsStoryFn<SolidRenderer, any>;
+        component?: infer Component;
+        args?: infer DefaultArgs;
+    }
+]
     ? Simplify<
-      (Component extends ComponentType<any>
-          ? ComponentProps<Component>
-          : unknown) &
-          ArgsFromMeta<SolidRenderer, TMetaOrCmpOrArgs>
+      (Component extends ComponentType<any> ? ComponentProps<Component> : unknown) &
+      ArgsFromMeta<SolidRenderer, TMetaOrCmpOrArgs>
     > extends infer TArgs
         ? StoryAnnotations<
             SolidRenderer,
-            TArgs,
-            SetOptional<
-                TArgs,
-          keyof TArgs & keyof (DefaultArgs & ActionArgs<TArgs>)
-            >
+            AddMocks<TArgs, DefaultArgs>,
+            SetOptional<TArgs, keyof TArgs & keyof DefaultArgs>
         >
         : never
-    : TMetaOrCmpOrArgs extends ComponentType<any>
+    : [TMetaOrCmpOrArgs] extends [ComponentType<any>]
         ? StoryAnnotations<SolidRenderer, ComponentProps<TMetaOrCmpOrArgs>>
         : StoryAnnotations<SolidRenderer, TMetaOrCmpOrArgs>;
 
-type ActionArgs<TArgs> = {
-    // This can be read as: filter TArgs on functions where we can assign a void function to that function.
-    // The docs addon argsEnhancers can only safely provide a default value for void functions.
-    // Other kind of required functions should be provided by the user.
-    [P in keyof TArgs as TArgs[P] extends (...args: any[]) => any
-        ? ((...args: any[]) => void) extends TArgs[P]
-            ? P
-            : never
-        : never]: TArgs[P];
-};
+// This performs a downcast to function types that are mocks, when a mock fn is given to meta args.
+export type AddMocks<TArgs, DefaultArgs> = Simplify<{
+    [T in keyof TArgs]: T extends keyof DefaultArgs
+        ? DefaultArgs[T] extends ((...args: any) => any) & { mock: object }
+            ? DefaultArgs[T]
+            : TArgs[T]
+        : TArgs[T];
+}>;
 
-export type Decorator<TArgs = StrictArgs> = DecoratorFunction<
-    SolidRenderer,
-    TArgs
->;
+export type Decorator<TArgs = StrictArgs> = DecoratorFunction<SolidRenderer, TArgs>;
 export type Loader<TArgs = StrictArgs> = LoaderFunction<SolidRenderer, TArgs>;
-export type StoryContext<TArgs = StrictArgs> = GenericStoryContext<
-    SolidRenderer,
-    TArgs
->;
+export type StoryContext<TArgs = StrictArgs> = GenericStoryContext<SolidRenderer, TArgs>;
 export type Preview = ProjectAnnotations<SolidRenderer>;