+++ packages/storybook-solid/src/types.ts
@@ -1,12 +1,14 @@
-import type { Args, WebRenderer } from '@storybook/types';
 import type { Component, JSXElement } from 'solid-js';
 
-export type { RenderContext } from '@storybook/types';
-export type { StoryContext } from '@storybook/types';
+import type { Args, Canvas, WebRenderer } from 'storybook/internal/types';
+
+export type { RenderContext, StoryContext } from '@storybook/types';
 
 export interface SolidRenderer extends WebRenderer {
+    // @ts-expect-error: Fix error in Github actions
     component: Component<this['T']>;
     storyResult: StoryFnSolidReturnType;
+    mount: (ui?: JSXElement) => Promise<Canvas>;
 }
 
 export interface ShowErrorArgs {
@@ -14,7 +16,12 @@ export interface ShowErrorArgs {
     description: string;
 }
 
-export type StoryFnSolidReturnType = JSXElement;
 export type ComponentsData = {
-    [key: string]: { args: Args; rendered?: boolean; disposeFn?: () => void };
+    [key: string]: {
+        args: Args;
+        rendered?: boolean;
+        disposeFn?: () => void;
+    };
 };
+
+export type StoryFnSolidReturnType = JSXElement;